+++ Raycoon.RayMigrator.Database.SqlServer/Templates/Repository_Environment_CheckInsert.sql
@@ -48,8 +48,6 @@ Note5 = "ResultCode catalog: see TemplateResultCode.cs in Shared project"
 ================================================================================
 */
 
--- Mandatory RepositoryVersion: DO NOT change manually, otherwise repository-inconsistencies may occur that results in migration errors !!!
-
 SET NOCOUNT ON;
 SET XACT_ABORT ON;